初步实现回复功能

+++ src/main/java/com/anglewang/community/dto/QuestionDTO.java
@@ -5,12 +5,12 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creatorId;
+    private Long creatorId;
     private Integer commentCount;
     private Integer viewCount;
     private Integer supportCount;
